change balance type

+++ src/main/kotlin/com/example/mtservice/controller/ProfileController.kt
@@ -31,7 +31,7 @@ class ProfileController(private val profileService: ProfileService) {
     fun newBalance(
         @AuthenticationPrincipal
         userDetails: CustomUserDetails?,
-        amount: Double,
+        amount: Long,
         model: Model,
     ): String {
         if (userDetails == null) {
