+++ src/main/kotlin/com/example/mtservice/dto/BalanceDto.kt
@@ -2,5 +2,5 @@ package com.example.mtservice.dto
 
 class BalanceDto(
     val id: Long,
-    val balance: Double,
+    val balance: Long,
 )
